+++ src/Traits/DetectsChanges.php
@@ -2,6 +2,7 @@
 
 namespace Spatie\Activitylog\Traits;
 
+use Illuminate\Support\Str;
 use Illuminate\Database\Eloquent\Model;
 use Spatie\Activitylog\Exceptions\CouldNotLogChanges;
 
@@ -27,24 +28,53 @@ trait DetectsChanges
     {
         $attributes = [];
 
-        if (isset(static::$logFillable)) {
-            $attributes = array_merge($attributes, $this->fillable);
+        if (isset(static::$logFillable) && static::$logFillable) {
+            $attributes = array_merge($attributes, $this->getFillable());
         }
 
-        if (isset(static::$logAttributes)) {
-            $attributes = array_merge($attributes, static::$logAttributes);
+        if ($this->shouldLogUnguarded()) {
+            $attributes = array_merge($attributes, array_diff(array_keys($this->getAttributes()), $this->getGuarded()));
+        }
+
+        if (isset(static::$logAttributes) && is_array(static::$logAttributes)) {
+            $attributes = array_merge($attributes, array_diff(static::$logAttributes, ['*']));
+
+            if (in_array('*', static::$logAttributes)) {
+                $attributes = array_merge($attributes, array_keys($this->getAttributes()));
+            }
+        }
+
+        if (isset(static::$logAttributesToIgnore) && is_array(static::$logAttributesToIgnore)) {
+            $attributes = array_diff($attributes, static::$logAttributesToIgnore);
         }
 
         return $attributes;
     }
 
-    public function shouldlogOnlyDirty(): bool
+    public function shouldLogOnlyDirty(): bool
     {
         if (! isset(static::$logOnlyDirty)) {
             return false;
         }
 
         return static::$logOnlyDirty;
+    }
+
+    public function shouldLogUnguarded(): bool
+    {
+        if (! isset(static::$logUnguarded)) {
+            return false;
+        }
+
+        if (! static::$logUnguarded) {
+            return false;
+        }
+
+        if (in_array('*', $this->getGuarded())) {
+            return false;
+        }
+
+        return true;
     }
 
     public function attributeValuesToBeLogged(string $processingEvent): array
@@ -79,11 +109,19 @@ trait DetectsChanges
     public static function logChanges(Model $model): array
     {
         $changes = [];
-        foreach ($model->attributesToBeLogged() as $attribute) {
-            if (str_contains($attribute, '.')) {
+        $attributes = $model->attributesToBeLogged();
+        $model = clone $model;
+        $model->append(array_filter($attributes, function ($key) use ($model) {
+            return ! array_key_exists($key, $model->getAttributes()) && $model->hasGetMutator($key);
+        }));
+        $model->setHidden(array_diff($model->getHidden(), $attributes));
+        $collection = collect($model);
+
+        foreach ($attributes as $attribute) {
+            if (Str::contains($attribute, '.')) {
                 $changes += self::getRelatedModelAttributeValue($model, $attribute);
             } else {
-                $changes += collect($model)->only($attribute)->toArray();
+                $changes += $collection->only($attribute)->toArray();
             }
         }
 
@@ -100,7 +138,7 @@ trait DetectsChanges
 
         $relatedModel = $model->$relatedModelName ?? $model->$relatedModelName();
 
-        return ["{$relatedModelName}.{$relatedAttribute}" => $relatedModel->$relatedAttribute];
+        return ["{$relatedModelName}.{$relatedAttribute}" => $relatedModel->$relatedAttribute ?? null];
     }
 
     private function array_diff_assoc_recursive($array1, $array2): array