+++ tests/Models/User.php
@@ -14,21 +14,13 @@ class User extends Model implements Authenticatable
 
     protected $guarded = [];
 
-    /**
-     * Get the name of the unique identifier for the user.
-     *
-     * @return string
-     */
+    protected $fillable = ['id', 'name'];
+
     public function getAuthIdentifierName()
     {
         return 'id';
     }
 
-    /**
-     * Get the unique identifier for the user.
-     *
-     * @return mixed
-     */
     public function getAuthIdentifier()
     {
         $name = $this->getAuthIdentifierName();
@@ -36,40 +28,20 @@ class User extends Model implements Authenticatable
         return $this->attributes[$name];
     }
 
-    /**
-     * Get the password for the user.
-     *
-     * @return string
-     */
     public function getAuthPassword()
     {
         return $this->attributes['password'];
     }
 
-    /**
-     * Get the token value for the "remember me" session.
-     *
-     * @return string
-     */
     public function getRememberToken()
     {
         return 'token';
     }
 
-    /**
-     * Set the token value for the "remember me" session.
-     *
-     * @param string $value
-     */
     public function setRememberToken($value)
     {
     }
 
-    /**
-     * Get the column name for the "remember me" token.
-     *
-     * @return string
-     */
     public function getRememberTokenName()
     {
         return 'tokenName';